~geolocalização, atualização e imagens do Imoveis

+++ airbnb-server/app/Controllers/Http/PropertyController.js
@@ -20,7 +20,12 @@ class PropertyController {
    * @param {View} ctx.view
    */
   async index({ request, response, view }) { // listar
-    const properties = Property.all()
+    const { latitude, longitude } = request.all()
+
+    const properties = Property.query()
+      .with('images')
+      .nearBy(latitude, longitude, 10)
+      .fetch()
 
     return properties
   }
@@ -33,7 +38,19 @@ class PropertyController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async store({ request, response }) {  // criar
+  async store({ auth, request, response }) {  // criar
+    const { id } = auth.user
+    const data = request.only([
+      'title',
+      'address',
+      'latitude',
+      'longitude',
+      'price'
+    ])
+
+    const property = await Property.create({ ...data, user_id: id })
+
+    return property
   }
 
   /**
@@ -62,6 +79,21 @@ class PropertyController {
    * @param {Response} ctx.response
    */
   async update({ params, request, response }) { // atualizar
+    const property = await Property.findOrFail(params.id)
+
+    const data = request.only([
+      'title',
+      'address',
+      'latitude',
+      'longitude',
+      'price'
+    ])
+
+    property.merge(data)
+
+    await property.save()
+
+    return property
   }
 
   /**
@@ -72,12 +104,12 @@ class PropertyController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async destroy({ params, request, response }) { // remover
+  async destroy({ auth, params, request, response }) { // remover
     const property = await Property.findOrFail(params.id)
 
-    // if (property.user_id !== auth.user.id) {
-    //   return response.status(401).send({ error: 'Not authorized' })
-    // }
+    if (property.user_id !== auth.user.id) {
+      return response.status(401).send({ error: 'Not authorized' })
+    }
 
     await property.delete()
   }
